change permissions for workspace chat settings
