+++ frontend/src/models/admin.js
@@ -139,42 +139,6 @@ const Admin = {
       });
   },
 
-  // Workspace Chats Mgmt
-  chats: async (offset = 0) => {
-    return await fetch(`${API_BASE}/admin/workspace-chats`, {
-      method: "POST",
-      headers: baseHeaders(),
-      body: JSON.stringify({ offset }),
-    })
-      .then((res) => res.json())
-      .catch((e) => {
-        console.error(e);
-        return [];
-      });
-  },
-  deleteChat: async (chatId) => {
-    return await fetch(`${API_BASE}/admin/workspace-chats/${chatId}`, {
-      method: "DELETE",
-      headers: baseHeaders(),
-    })
-      .then((res) => res.json())
-      .catch((e) => {
-        console.error(e);
-        return { success: false, error: e.message };
-      });
-  },
-  exportChats: async () => {
-    return await fetch(`${API_BASE}/admin/export-chats`, {
-      method: "GET",
-      headers: baseHeaders(),
-    })
-      .then((res) => res.text())
-      .catch((e) => {
-        console.error(e);
-        return null;
-      });
-  },
-
   // System Preferences
   systemPreferences: async () => {
     return await fetch(`${API_BASE}/admin/system-preferences`, {
